Module Raw_grammar.Atom

A grammatical type which classifies atoms.

type t =

| String	(*
Any atom.
*)

| Bool	(*
One of true, false, True, or False.
*)

| Char	(*
A single-character atom.
*)

| Float	(*
An atom which parses as a float.
*)

| Int	(*
An atom which parses as an integer, such as int or int64.
*)

| This of {

ignore_capitalization : bool;

string : string;

}	(*
Exactly that string, possibly modulo case in the first character.
*)